only admin can update and delete stock

+++ server/controllers/stocks.controller.js
@@ -1,6 +1,6 @@
 import Stock from '../models/stock';
-import {AppError} from '../utilities/appError';
-import {AppResponse} from '../utilities/appResponse';
+import { AppError } from '../utilities/appError';
+import { AppResponse } from '../utilities/appResponse';
 import DataRepository from '../data_access/dataRepository';
 
 const repo = new DataRepository();
@@ -8,7 +8,7 @@ const repo = new DataRepository();
 export function list(req, res, next) {
     repo.getAll(Stock)
         .then(results => {
-            if(!results) {
+            if (!results) {
                 return next(new AppError('No stocks found', 404));
             }
 
@@ -17,8 +17,8 @@ export function list(req, res, next) {
         .catch(err => next(err));
 }
 
-export function create(req, res, next){
-    repo.nextStockSeqNr().then(function(seqNr){
+export function create(req, res, next) {
+    repo.nextStockSeqNr().then(function (seqNr) {
         const stock = new Stock({
             seqNr: seqNr,
             companyName: req.body.name.toUpperCase(),
@@ -28,31 +28,43 @@ export function create(req, res, next){
             investableInstrument: req.body.investableInstrument,
             businessDescription: req.body.businessDescription
         });
-    
+
         repo.save(stock)
             .then(() => res.status(201).json(new AppResponse('Stock added', stock)))
-            .catch(err => next(err)); 
-    });    
-}
-
-export function update(req,res, next){
-    const stock = new Stock({
-        _id: req.body._id,
-        companyName: req.body.name.toUpperCase(),
-        country: req.body.country.toUpperCase(),
-        website: req.body.website.toLowerCase(),
-        addedBy: res.locals.user,
-        investableInstrument: req.body.investableInstrument,
-        businessDescription: req.body.businessDescription
+            .catch(err => next(err));
     });
-    repo.update(Stock, {"_id": stock._id}, stock)
-        .then(()=> res.status(201).json(new AppResponse("Stock Updated", stock)))
-        .catch(err=>next(err));
 }
 
-export function remove(req,res, next){
-    let id = (req.params.stockId);
-    repo.deleteStockById(id)
-        .then(() => res.status(201).json(new AppResponse("Stock Deleted", id)))
-        .catch(err => next(err));
+export function update(req, res, next) {
+    let isCurrentLoggedInUserAdmin = res.locals.user.isAdmin == true ? true : false;
+    if (isCurrentLoggedInUserAdmin) {
+        const stock = new Stock({
+            _id: req.body._id,
+            companyName: req.body.name.toUpperCase(),
+            country: req.body.country.toUpperCase(),
+            website: req.body.website.toLowerCase(),
+            addedBy: res.locals.user,
+            investableInstrument: req.body.investableInstrument,
+            businessDescription: req.body.businessDescription
+        });
+        repo.update(Stock, { "_id": stock._id }, stock)
+            .then(() => res.status(201).json(new AppResponse("Stock Updated", stock)))
+            .catch(err => next(err));
+    }
+    else {
+        return res.status(400).json(new AppResponse('Access denied ', null));
+    }
+}
+
+export function remove(req, res, next) {
+    let isCurrentLoggedInUserAdmin = res.locals.user.isAdmin == true ? true : false;
+    if (isCurrentLoggedInUserAdmin) {
+        let id = (req.params.stockId);
+        repo.deleteStockById(id)
+            .then(() => res.status(201).json(new AppResponse("Stock Deleted", id)))
+            .catch(err => next(err));
+    }
+    else {
+        return res.status(400).json(new AppResponse('Access denied ', null));
+    }
 }
